+++ src/main/java/org/mbari/mxm/ProviderManager.java
@@ -1,18 +1,15 @@
 package org.mbari.mxm;
 
 import java.time.OffsetDateTime;
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.mbari.mxm.db.argument.Argument;
 import org.mbari.mxm.db.argument.ArgumentService;
-import org.mbari.mxm.db.asset.Asset;
-import org.mbari.mxm.db.asset.AssetService;
-import org.mbari.mxm.db.assetClass.AssetClass;
-import org.mbari.mxm.db.assetClass.AssetClassService;
 import org.mbari.mxm.db.mission.Mission;
 import org.mbari.mxm.db.mission.MissionService;
 import org.mbari.mxm.db.mission.MissionStatusType;
@@ -24,23 +21,17 @@ import org.mbari.mxm.db.parameter.Parameter;
 import org.mbari.mxm.db.parameter.ParameterService;
 import org.mbari.mxm.db.provider.Provider;
 import org.mbari.mxm.db.provider.ProviderApiType;
-import org.mbari.mxm.db.unit.Unit;
-import org.mbari.mxm.db.unit.UnitService;
 import org.mbari.mxm.graphql.ProviderPingException;
 import org.mbari.mxm.provider_client.MxmProviderClient;
 import org.mbari.mxm.provider_client.MxmProviderClientBuilder;
-import org.mbari.mxm.provider_client.responses.*;
+import org.mbari.mxm.provider_client.responses.MissionTemplateResponse;
+import org.mbari.mxm.provider_client.responses.MissionValidationResponse;
+import org.mbari.mxm.provider_client.responses.PingResponse;
 import org.mbari.mxm.provider_client.rest.PostMissionPayload;
 
 @ApplicationScoped
 @Slf4j
 public class ProviderManager {
-
-  @Inject AssetClassService assetClassService;
-
-  @Inject AssetService assetService;
-
-  @Inject UnitService unitService;
 
   @Inject MissionTemplateService missionTemplateService;
 
@@ -109,49 +100,7 @@ public class ProviderManager {
     public void postInsertProvider(Provider provider) {
       log.debug("postInsertProvider: provider={}", provider);
 
-      // Asset Classes
-      var assetClasses = mxmProviderClient.getAssetClasses();
-      log.debug("postInsertProvider: assetClasses=>{}", assetClasses);
-      createAssetClasses(provider, assetClasses.result);
-
-      // Units:
-      if (provider.usesUnits) {
-        getAndCreateUnits(provider);
-      }
-
-      // MissionTpls:
       getAndCreateMissionTplsForDirectory(provider, "/");
-    }
-
-    private void createAssetClasses(
-        Provider provider, List<AssetClassesResponse.AssetClass> assetClasses) {
-      assetClasses.forEach(assetClass -> createAssetClass(provider, assetClass));
-    }
-
-    private void createAssetClass(Provider provider, AssetClassesResponse.AssetClass assetClass) {
-      log.debug("createAssetClass: assetClass=>{}", assetClass);
-      var ac =
-          assetClassService.createAssetClass(
-              new AssetClass(
-                  provider.providerId, assetClass.assetClassName, assetClass.description));
-      log.debug("createAssetClass: created=>{}", ac);
-
-      createAssets(provider, assetClass);
-    }
-
-    private void createAssets(Provider provider, AssetClassesResponse.AssetClass assetClass) {
-      assetClass.assets.forEach(asset -> createAsset(provider, assetClass, asset));
-    }
-
-    private void createAsset(
-        Provider provider,
-        AssetClassesResponse.AssetClass assetClass,
-        AssetClassesResponse.Asset a) {
-      log.debug("createAsset: a=>{}", a);
-      var asset = new Asset(provider.providerId, a.assetId);
-      asset.className = assetClass.assetClassName;
-      // TODO capture description in AssetClassesResponse.Asset
-      assetService.createAsset(asset);
     }
 
     private void getAndCreateMissionTplsForDirectory(Provider provider, String directory) {
@@ -263,25 +212,6 @@ public class ProviderManager {
             parameter.description = pp.description;
             parameterService.createParameter(parameter);
           });
-    }
-
-    private void getAndCreateUnits(Provider provider) {
-      var units =
-          mxmProviderClient.getUnits().result.stream()
-              .collect(Collectors.partitioningBy(u -> u.baseUnit == null));
-
-      // first, create units without a base:
-      units.get(true).forEach(u -> createUnit(provider, u));
-      // then the others:
-      units.get(false).forEach(u -> createUnit(provider, u));
-    }
-
-    private void createUnit(Provider provider, UnitsResponse.Unit unit) {
-      var u = new Unit(provider.providerId, unit.name);
-      u.abbreviation = unit.abbreviation;
-      u.baseUnit = unit.baseUnit;
-      var created = unitService.createUnit(u);
-      log.trace("createUnit: created=>{}", created);
     }
 
     /** Refreshes mission template information from the provider. */