+++ src/main/java/org/mbari/mxm/graphql/ForAsset.java
@@ -1,7 +1,5 @@
 package org.mbari.mxm.graphql;
 
-import static java.util.stream.Collectors.toList;
-
 import io.quarkus.arc.Unremovable;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import java.util.List;
@@ -11,8 +9,6 @@ import org.eclipse.microprofile.graphql.Source;
 import org.mbari.mxm.db.asset.Asset;
 import org.mbari.mxm.db.assetClass.AssetClass;
 import org.mbari.mxm.db.assetClass.AssetClassService;
-import org.mbari.mxm.db.provider.Provider;
-import org.mbari.mxm.db.provider.ProviderService;
 
 @ApplicationScoped
 @RegisterForReflection
@@ -20,14 +16,7 @@ import org.mbari.mxm.db.provider.ProviderService;
 public class ForAsset {
   @Inject AssetClassService assetClassService;
 
-  @Inject ProviderService providerService;
-
   public List<AssetClass> assetClass(@Source List<Asset> assets) {
     return assetClassService.getAssetClasses(assets);
-  }
-
-  public List<Provider> provider(@Source List<Asset> assets) {
-    List<String> providerIds = assets.stream().map(e -> e.providerId).collect(toList());
-    return providerService.getProviders(providerIds);
   }
 }