+++ src/main/java/org/mbari/mxm/graphql/ForAssetClass.java
@@ -1,7 +1,5 @@
 package org.mbari.mxm.graphql;
 
-import static java.util.stream.Collectors.toList;
-
 import io.quarkus.arc.Unremovable;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import java.util.List;
@@ -11,8 +9,6 @@ import org.eclipse.microprofile.graphql.Source;
 import org.mbari.mxm.db.asset.Asset;
 import org.mbari.mxm.db.asset.AssetService;
 import org.mbari.mxm.db.assetClass.AssetClass;
-import org.mbari.mxm.db.provider.Provider;
-import org.mbari.mxm.db.provider.ProviderService;
 
 @ApplicationScoped
 @RegisterForReflection
@@ -20,14 +16,7 @@ import org.mbari.mxm.db.provider.ProviderService;
 public class ForAssetClass {
   @Inject AssetService assetService;
 
-  @Inject ProviderService providerService;
-
   public List<List<Asset>> assets(@Source List<AssetClass> assetClasses) {
     return assetService.getAssetsMultipleForAssetClasses(assetClasses);
-  }
-
-  public List<Provider> provider(@Source List<AssetClass> assetClasses) {
-    List<String> providerIds = assetClasses.stream().map(e -> e.providerId).collect(toList());
-    return providerService.getProviders(providerIds);
   }
 }