+++ ChangeLog.md
@@ -1,3 +1,15 @@
+2022-10
+
+- Model revision: 
+  - `Unit` model now by itself
+    - So, the set of units of measure is to be a common resource
+    - Initially to be populated from the LRAUV/TethysDash system
+  - `AssetClass` and `Asset` models now by themselves
+    - As before, provider continues to indicate the `AssetClass`es used by its mission templates.
+    - This is in preparation for eventual integration of the TrackingDB information *or* a similar common
+      place that defines the asset classes (types) and assets (platforms) referenceable by any provider.
+    - for convenience, adding some asset classes and assets at database setup time.
+
 2022-09
 
 - adding mission validation